nic: add nic
